+++ main1049.cpp
@@ -142,7 +142,7 @@ blHWw = G8_KwSe(SJ3T3_, (short)-1329343881);
 }
 else if (d4Y1Zm7)
 {
-Mbo3n = G8_KwSe(SJ3T3_, -405724126) % 1422509334;
+Mbo3n = G8_KwSe(SJ3T3_, (short)-405724126) % 1422509334;
 }
 else if (OAf_69XO8 != 1877638358)
 {
